feat: separated project on single function
