+++ src/controllers/getUser.ts
@@ -1,19 +1,20 @@
 import { ServerResponse } from 'http';
+import { writeStatus404 } from '../utils/status/writeStatus404';
+import { writeStatus400 } from '../utils/status/writeStatus400';
 import { valid } from '../utils/validate';
 import { findUserById } from '../models/userModel';
+import { writeStatus200 } from '../utils/status/writeStatus200';
 
 export const getUser = async (res:ServerResponse, id: string) => {
-  const user = await findUserById(id);
   try {
     if (!valid(id)) {
-      res.writeHead(400, { 'Content-Type': 'application/json' });
-      res.end(JSON.stringify('Bad Request'));
-    } else if (!user) {
-      res.writeHead(404, { 'Content-Type': 'application/json' });
-      res.end(JSON.stringify('User Not Found'));
+      writeStatus400(res);
+    }
+    const user = await findUserById(id);
+    if (!user) {
+      writeStatus404(res);
     } else {
-      res.writeHead(200, { 'Content-Type': 'application/json' });
-      res.end(JSON.stringify(user));
+      writeStatus200(res, user);
     }
   } catch (e) {
     throw new Error('Operation failed');
