+++ src/controllers/getUsers.ts
@@ -1,12 +1,12 @@
 import { ServerResponse } from 'http';
+import { writeStatus200 } from '../utils/status/writeStatus200';
 
 import { findAllUsers } from '../models/userModel';
 
 export const getUsers = async (res:ServerResponse) => {
   try {
     const users = await findAllUsers();
-    res.writeHead(200, { 'Content-Type': 'application/json' });
-    res.end(JSON.stringify(users));
+    writeStatus200(res, users);
   } catch (e) {
     throw new Error('Operation failed');
   }
