+++ src/utils/status/writeStatus400.ts
@@ -1,0 +1,6 @@
+import { ServerResponse } from 'http';
+
+export const writeStatus400 = (res:ServerResponse) => {
+  res.writeHead(400, { 'Content-Type': 'application/json' });
+  return res.end(JSON.stringify('Bad Request'));
+};
