+++ src/utils/status/writeStatus404.ts
@@ -1,0 +1,6 @@
+import { ServerResponse } from 'http';
+
+export const writeStatus404 = (res:ServerResponse) => {
+  res.writeHead(404, { 'Content-Type': 'application/json' });
+  res.end(JSON.stringify('User Not Found'));
+};
